% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zysno.R
\name{errors_item_pair}
\alias{errors_item_pair}
\title{Calculate errors for one item pair}
\usage{
errors_item_pair(item_pair)
}
\arguments{
\item{item_pair}{2-column df or matrix}
}
\value{
number of errors
}
\description{
Calculate errors for one item pair
}
