% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zysno.R
\name{get_errors_for_one_item_pair}
\alias{get_errors_for_one_item_pair}
\title{Calculate errors for one item pair (old version)}
\usage{
get_errors_for_one_item_pair(item_pair)
}
\arguments{
\item{item_pair}{2-column df or matrix}
}
\value{
number of errors
}
\description{
Uses Kronecker product for generating combinations
is relatively slow
}
