% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zysno.R
\name{get_expected_errors}
\alias{get_expected_errors}
\title{Calculate number of expected errors for one item pair}
\usage{
get_expected_errors(item_pair)
}
\value{
number of expected errors
}
\description{
Calculate number of expected errors for one item pair
}
