% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zysno.R
\name{zysnotize}
\alias{zysnotize}
\title{Main function of package zysno}
\usage{
zysnotize(d)
}
\arguments{
\item{d}{dataset, rows are participants, columns are items}
}
\value{
list of different outputs
}
\description{
Performs analysis of unidimensionality based on the paper by Zysno
}
